move dialogs into separate functionality
